refactor(media-refresh): split into modules
